Add diff "kinds"
The purpose of diff "kinds" is to distinguish how diffs should be
represented independent of the type of the diff segments. This lets you
format string diffs differently from a diff of slices of strings.

This is implemented as a type parameter instead of a value parameter so
that implementations of the formatter can be generic over the kind.

I prefer this over using the type of the collection you're diffing
because different collection types might have the same kind (e.g. str vs
String).